requirements.txt
aiofiles==25.1.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
bcrypt==4.0.1
billiard==4.2.2
black==25.11.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.123.10
fastapi-cache2==0.2.2
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
kombu==5.6.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
PyYAML==6.0.3
redis==7.1.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
ruff==0.14.14
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
